refactor: reorganize Signup components and implement WorkspaceGuestSignUpView
