+++ src/views/Signup/WorkspaceGuestSignUpView.test.tsx
@@ -1,14 +1,14 @@
 import { beforeEach, beforeAll, describe, expect, it, vi, Mock } from 'vitest';
 import { screen, fireEvent, render } from '@testing-library/react';
-import WorkspaceGuestSingUpView from './WorkspaceGuestSignUp';
-import { userActions } from '../../../app/store/slices/user';
-import * as keysService from '../../../app/crypto/services/keys.service';
-import { encryptTextWithKey } from '../../../app/crypto/services/utils';
+import WorkspaceGuestSingUpView from './WorkspaceGuestSignUpView';
+import { userActions } from '../../app/store/slices/user';
+import * as keysService from '../../app/crypto/services/keys.service';
+import { encryptTextWithKey } from '../../app/crypto/services/utils';
 import { UserSettings } from '@internxt/sdk/dist/shared/types/userSettings';
-import { useSignUp } from '../hooks/useSignup';
+import { useSignUp } from './hooks/useSignup';
 import { Buffer } from 'node:buffer';
 import { generateMnemonic } from 'bip39';
-import envService from '../../../app/core/services/env.service';
+import envService from '../../app/core/services/env.service';
 
 const mockSecret = '123456789QWERTY';
 const mockMagicIv = '12345678912345678912345678912345';
@@ -93,7 +93,7 @@ describe('onSubmit', () => {
       X: () => <div>Mocked X Icon</div>,
     }));
 
-    vi.mock('../../../app/auth/components/PasswordInput/PasswordInput', () => {
+    vi.mock('../../app/auth/components/PasswordInput/PasswordInput', () => {
       return {
         __esModule: true,
         default: vi.fn(({ register, ...props }) => (
@@ -110,25 +110,25 @@ describe('onSubmit', () => {
       };
     });
 
-    vi.mock('../hooks/useSignup', () => ({
+    vi.mock('./hooks/useSignup', () => ({
       useSignUp: vi.fn().mockReturnValue({ doRegisterPreCreatedUser: vi.fn() }),
       parseUserSettingsEnsureKyberKeysAdded: vi.importActual,
     }));
 
-    vi.mock('../../../app/shared/components/PasswordStrengthIndicator', () => ({
+    vi.mock('../../app/shared/components/PasswordStrengthIndicator', () => ({
       default: {
         PasswordStrengthIndicator: () => <div>Mocked Password Strength Indicator</div>,
       },
     }));
 
-    vi.mock('../../../app/core/services/error.service', () => ({
+    vi.mock('../../app/core/services/error.service', () => ({
       default: {
         castError: vi.fn().mockImplementation((e) => ({ message: e.message || 'Default error message' })),
         reportError: vi.fn(),
       },
     }));
 
-    vi.mock('../../../app/core/services/local-storage.service', () => ({
+    vi.mock('../../app/core/services/local-storage.service', () => ({
       default: {
         get: vi.fn(),
         clear: vi.fn(),
@@ -137,7 +137,7 @@ describe('onSubmit', () => {
       },
     }));
 
-    vi.mock('../../../app/core/services/navigation.service', () => ({
+    vi.mock('../../app/core/services/navigation.service', () => ({
       default: {
         push: vi.fn(),
         history: {
@@ -148,7 +148,7 @@ describe('onSubmit', () => {
       },
     }));
 
-    vi.mock('../../../app/i18n/provider/TranslationProvider', () => ({
+    vi.mock('../../app/i18n/provider/TranslationProvider', () => ({
       useTranslationContext: vi.fn().mockReturnValue({
         translate: vi.fn().mockImplementation((value: string) => {
           return value;
@@ -156,7 +156,7 @@ describe('onSubmit', () => {
       }),
     }));
 
-    vi.mock('../../../app/shared/views/ExpiredLink/ExpiredLinkView', () => ({
+    vi.mock('../../app/shared/views/ExpiredLink/ExpiredLinkView', () => ({
       default: {
         ExpiredLink: vi.fn(),
       },
@@ -171,11 +171,11 @@ describe('onSubmit', () => {
       useDispatch: vi.fn(() => vi.fn()),
     }));
 
-    vi.mock('../utils', () => ({
+    vi.mock('./utils', () => ({
       onChangePasswordHandler: vi.fn(),
     }));
 
-    vi.mock('../../../app/core/services/workspace.service', () => ({
+    vi.mock('../../app/core/services/workspace.service', () => ({
       default: {
         validateWorkspaceInvitation: vi.fn().mockImplementation(() => {
           return true;
@@ -183,29 +183,29 @@ describe('onSubmit', () => {
       },
     }));
 
-    vi.mock('../../../app/store/hooks', () => ({
+    vi.mock('../../app/store/hooks', () => ({
       useAppDispatch: vi.fn().mockReturnValue(vi.fn()),
     }));
 
-    vi.mock('../../../app/store/slices/plan', () => ({
+    vi.mock('../../app/store/slices/plan', () => ({
       planThunks: {
         initializeThunk: vi.fn(),
       },
     }));
 
-    vi.mock('../../../app/store/slices/products', () => ({
+    vi.mock('../../app/store/slices/products', () => ({
       productsThunks: {
         initializeThunk: vi.fn(),
       },
     }));
 
-    vi.mock('../../../app/store/slices/referrals', () => ({
+    vi.mock('../../app/store/slices/referrals', () => ({
       referralsThunks: {
         initializeThunk: vi.fn(),
       },
     }));
 
-    vi.mock('../../../app/store/slices/user', () => ({
+    vi.mock('../../app/store/slices/user', () => ({
       initializeUserThunk: vi.fn(),
       userActions: {
         setUser: vi.fn(),
