+++ src/views/Signup/WorkspaceGuestSignUpView.tsx
@@ -1,21 +1,21 @@
 import { auth } from '@internxt/lib';
-import { useSignUp } from '../hooks/useSignup';
-import { useGuestSignupState } from '../hooks/useGuestSignupState';
-import { useInvitationValidation } from '../hooks/useInvitationValidation';
-import { useGuestSignupForm } from '../hooks/useGuestSignupForm';
-import navigationService from '../../../app/core/services/navigation.service';
-import { AppView, IFormValues } from '../../../app/core/types';
-import { useTranslationContext } from '../../../app/i18n/provider/TranslationProvider';
-import ExpiredLink from '../../../app/shared/views/ExpiredLink/ExpiredLinkView';
-import { useAppDispatch } from '../../../app/store/hooks';
-import { userActions } from '../../../app/store/slices/user';
+import { useSignUp } from './hooks/useSignup';
+import { useGuestSignupState } from './hooks/useGuestSignupState';
+import { useInvitationValidation } from './hooks/useInvitationValidation';
+import { useGuestSignupForm } from './hooks/useGuestSignupForm';
+import navigationService from '../../app/core/services/navigation.service';
+import { AppView, IFormValues } from '../../app/core/types';
+import { useTranslationContext } from '../../app/i18n/provider/TranslationProvider';
+import ExpiredLink from '../../app/shared/views/ExpiredLink/ExpiredLinkView';
+import { useAppDispatch } from '../../app/store/hooks';
+import { userActions } from '../../app/store/slices/user';
 import queryString from 'query-string';
 import { useEffect } from 'react';
 import { SubmitHandler, useForm, useWatch } from 'react-hook-form';
-import { onChangePasswordHandler } from '../utils';
-import CreateAccountForm from './CreateAccountForm';
-import workspacesService from '../../../app/core/services/workspace.service';
-import { guestSignupOnSubmit } from '../utils/guestSignupOnSubmit';
+import { onChangePasswordHandler } from './utils';
+import CreateAccountForm from './components/CreateAccountForm';
+import workspacesService from '../../app/core/services/workspace.service';
+import { guestSignupOnSubmit } from './utils/guestSignupOnSubmit';
 
 function WorkspaceGuestSingUpView(): JSX.Element {
   const dispatch = useAppDispatch();
