+++ src/views/Signup/components/index.ts
@@ -1,4 +1,3 @@
 export { default as SignupForm } from './SignupForm';
 export { SignupBlog, AuthView, ButtonAuth, SignupComponent } from './SignupBlog';
 export { default as CreateAccountForm } from './CreateAccountForm';
-export { default as WorkspaceGuestSignUp } from './WorkspaceGuestSignUp';
