+++ src/views/Signup/index.ts
@@ -1,2 +1,3 @@
 export { default } from './SignupView';
 export { default as ShareGuestSignUpView } from './ShareGuestSignUpView';
+export { default as WorkspaceGuestSignUpView } from './WorkspaceGuestSignUpView';
